feat: rework overview
